refactor further to separate tests logically
